refactor(website): simplify login error handling
