+++ .impeccable/critique/2026-07-26T11-52-31Z__website-app-routes-login-tsx.md
@@ -1,0 +1,100 @@
+---
+target: website/app/routes/login.tsx
+total_score: 24
+max_score: 40
+na_heuristics: 
+p0_count: 0
+p1_count: 2
+timestamp: 2026-07-26T11-52-31Z
+slug: website-app-routes-login-tsx
+---
+# Fontsource login critique
+
+## Design Health Score
+
+| Heuristic | Score | Summary |
+|---|---:|---|
+| Visibility of system status | 3/4 | Provider loading, mutual disabling, and a failure alert are implemented. Success is communicated only through redirect. |
+| Match with the real world | 3/4 | Familiar provider names and icons, but the page does not explain the real-world purpose of a Fontsource account. |
+| User control and freedom | 2/4 | Global navigation provides an exit, but the optional nature and current scope of the account are unclear. |
+| Consistency and standards | 3/4 | Mantine, IBM Plex, compact radii, and bordered surfaces fit the product system. |
+| Error prevention | 2/4 | Duplicate submission is prevented, but the OAuth trust and data-use decision is unexplained. |
+| Recognition rather than recall | 3/4 | Actions are explicit, while account benefits and collection behavior must be inferred. |
+| Flexibility and efficiency | 2/4 | Two low-friction provider paths exist, with no alternate method or explicit guest path. |
+| Aesthetic and minimalist design | 3/4 | Calm and focused, but so neutral that the surface feels category-interchangeable. |
+| Error recovery | 2/4 | Recovery copy is actionable but does not retain provider or safe error context. |
+| Help and documentation | 1/4 | No contextual privacy, account-purpose, or sign-in help appears at the trust decision. |
+| **Total** | **24/40** | **Acceptable foundation with significant clarity and trust gaps.** |
+
+## Design Specificity Verdict
+
+**Under-authored and category-interchangeable.** The centered social-login card is coherent with Fontsource's light-first palette, neutral typography, compact geometry, and bordered surfaces, but it could belong to nearly any developer product unchanged. Fontsource appears in the heading and global shell, while the product's distinctive values and the truthful reason for identity are absent.
+
+The deterministic Impeccable scan returned no rule violations. That means the implementation avoids common mechanical design anti-patterns, not that the product decision is complete. Browser overlays and rendered inspection were unavailable because the connected browser runtime exposed no browser backend.
+
+## Overall Impression
+
+The page is cognitively light and operationally clear: users see one heading and two recognizable OAuth actions. Its weakness is not clutter but missing context. Immediately before leaving Fontsource for an identity provider, users cannot tell what signing in unlocks, what information is used, or what useful destination awaits them afterward.
+
+## What Is Working
+
+- One clear task with two fully labeled provider actions.
+- Native buttons, a real `h1`, decorative icons hidden from assistive technology, and a semantic error alert.
+- Provider-specific loading and mutual disabling prevent accidental duplicate submissions.
+- The restrained card, spacing, and typography fit Fontsource's existing visual system.
+- The deterministic detector reported zero findings.
+
+## Priority Issues
+
+### [P1] The account has no stated product job
+
+The route says how to authenticate but not why. With collections currently described as browser-local and the signed-in destination acting only as proof, users may infer sync, publishing, or required membership that does not yet exist.
+
+**Fix:** Keep the hidden utility route, but add one short, strictly truthful line that names the current purpose. Do not promise collection sync or publishing until those workflows exist.
+
+**Suggested command:** `$impeccable shape`
+
+### [P1] The OAuth trust decision lacks concise reassurance
+
+Fontsource's self-hosting and privacy positioning raises the standard for an unexplained Google or GitHub handoff. The page does not say what identity information is used or point to relevant privacy terms.
+
+**Fix:** Add one restrained data-use sentence and a nearby Privacy Policy link. Keep it factual and avoid the previously rejected generic “no password required” or “browsing does not require an account” copy.
+
+**Suggested command:** `$impeccable clarify`
+
+### [P2] Error recovery loses provider and failure context
+
+“Couldn't log you in” and “Please try again” cover every failure identically. A cancellation, provider rejection, network problem, and local configuration issue lead to different next actions.
+
+**Fix:** Retain the attempted provider and map safe error categories to specific messages. Offer explicit retry and alternate-provider actions where the provider response permits it.
+
+**Suggested command:** `$impeccable harden`
+
+### [P2] The surface is visually compliant but anonymous
+
+The default bordered card and equal neutral buttons are appropriate but have little Fontsource-specific authorship. Violet does not establish a meaningful action voice, and the empty space beneath the title amplifies the missing context.
+
+**Fix:** Resolve the account purpose first. Then add one restrained signature such as a small utility eyebrow and clearer text hierarchy. Keep providers visually equal unless product evidence supports recommending one.
+
+**Suggested command:** `$impeccable bolder`
+
+## Persona Red Flags
+
+- **Jordan, confused first-timer:** can identify the next click but cannot predict whether login is required, whether collections sync, or what happens after OAuth.
+- **Sam, accessibility-dependent user:** benefits from solid semantics, but generic recovery requires remembering which provider was attempted. Live focus, loading announcements, and contrast remain unverified.
+- **Casey, distracted mobile user:** gets a low-input flow, but the 42px controls are slightly below a 44px touch target and external OAuth has no reassurance about return state.
+- **Mika, privacy-conscious developer/designer:** is likely to question why a self-hosting-oriented product needs Google or GitHub identity and what provider data is retained.
+
+## Minor Observations
+
+- A short purpose sentence would repair both the card's top-heavy composition and its trust gap.
+- The full global shell makes this hidden route feel like an orphaned content page rather than a deliberate account boundary.
+- The account destination confirms identity but provides no meaningful peak-end payoff.
+- The callback returns through `/login` before the loader redirects to `/account`; rendered authentication should be checked for visible flicker when a browser backend is available.
+
+## Questions to Consider
+
+1. Is the account's current truthful purpose simply proving creator identity, or should the page stay effectively internal until submissions exist?
+2. What minimum provider data is stored, and which existing privacy page can state that clearly?
+3. Should Google and GitHub remain equal, or does the creator workflow justify recommending GitHub later?
+4. What should `/account` offer first so a successful login ends with useful progress rather than only confirmation?
