+++ .impeccable/hook.cache.json
@@ -1,0 +1,1 @@
+{"version":1,"sessions":{"019f7aea-4d61-7593-96ff-96966aaf1eb4":{"updatedAt":1785068230426,"files":{"/Users/ayu.master/.codex/worktrees/f644/fontsource/website/app/utils/login-error.ts":{"editCount":5,"findings":[]},"/Users/ayu.master/.codex/worktrees/f644/fontsource/website/tests/login-error.test.ts":{"editCount":5,"findings":[]},"/Users/ayu.master/.codex/worktrees/f644/fontsource/website/app/routes/login.tsx":{"editCount":6,"findings":[],"cleanAcked":true},"/Users/ayu.master/.codex/worktrees/f644/fontsource/website/app/utils/auth.server.ts":{"editCount":1,"findings":[]},"/Users/ayu.master/.codex/worktrees/f644/fontsource/website/tests/auth.worker.ts":{"editCount":1,"findings":[]},"/Users/ayu.master/.codex/worktrees/f644/fontsource/website/app/routes/account.tsx":{"editCount":1,"findings":[]}}}}}
